category/games/2026/02/20260210-043051-vector-super-mario-bros-jump-platform/pyproject.toml
[project]
name = "vector-super-mario-bros-jump-platform"
version = "0.1.0"
description = "Navigate moving platforms to reach the goal in this physics-based horizontal platformer"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pygame-ce>=2.5.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
